config-loader: split up config schema implementation and move into separate lib
