+++ packages/config-loader/src/lib/schema/types.ts
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2020 Spotify AB
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { AppConfig } from '@backstage/config';
+import { JSONSchema7 as JSONSchema } from 'json-schema';
+
+export type ConfigSchemaPackageEntry = {
+  value: JSONSchema;
+  path: string;
+};
+
+export const CONFIG_VISIBILITIES = ['frontend', 'backend', 'secret'] as const;
+
+export type ConfigVisibility = typeof CONFIG_VISIBILITIES[number];
+
+export const DEFAULT_CONFIG_VISIBILITY: ConfigVisibility = 'backend';
+
+type ValidationError = string;
+
+type ValidationResult = {
+  errors?: ValidationError[];
+  visibilityByPath: Map<string, ConfigVisibility>;
+};
+
+export type ValidationFunc = (configs: AppConfig[]) => ValidationResult;
+
+type ConfigProcessingOptions = {
+  visibilities?: ConfigVisibility[];
+};
+
+export type ConfigSchema = {
+  process(
+    appConfigs: AppConfig[],
+    options?: ConfigProcessingOptions,
+  ): AppConfig[];
+};
